Changed modules name & 1.0.1442.BUILD-SNAPSHOT
Changed the name of API modules to :: API (instead of JAXRS API)
Changed the name of JAXRS IMPL to :: IMPL JAXRS so other IMPL will come
later and will be aligned.
Changed the Versioning Number according to the Spring Versioning
Numbering that seems to me an effective way of numbering the code.